requirements.txt
streamlit~=1.39.0
audio_recorder_streamlit
google-generativeai

SpeechRecognition~=3.10.4
pydub~=0.25.1
protobuf~=5.28.2